+++ packages/core/src/utils.ts
@@ -24,6 +24,9 @@ export const isString = (val: unknown): val is string =>
 export const isSymbol = (val: unknown): val is symbol =>
   typeof val === "symbol";
 
+export const isBoolean = (val: unknown): val is Boolean =>
+  typeof val === "boolean";
+
 export const isDate = (val: unknown): val is Date =>
   isObject(val) && isFunction(val.getTime);
 
@@ -39,6 +42,9 @@ export const isElement = (val: unknown): val is Element =>
 export function isPromise<T = any>(val: unknown): val is Promise<T> {
   return isObject(val) && isFunction(val.then) && isFunction(val.catch);
 }
+
+export const NO_OP = () => { };
+export const FALSE_OP = () => false;
 
 export function promisedTimeout(timeout: number): Promise<void> {
   return new Promise(res => {