+++ packages/web/src/event/event.ts
@@ -3,6 +3,38 @@ import { RefTyped, wrap } from "@vue-composable/core";
 
 export type RemoveEventFunction = () => void;
 
+export function useEvent<
+  T extends {
+    addEventListener: (
+      name: string,
+      listener: EventListenerOrEventListenerObject
+    ) => any;
+    removeEventListener: Function;
+  },
+  M,
+  K extends keyof M
+>(
+  el: RefTyped<T>,
+  name: K,
+  listener: (this: T, ev: M[K]) => any
+): RemoveEventFunction;
+export function useEvent<
+  T extends {
+    addEventListener: (
+      name: string,
+      listener: EventListenerOrEventListenerObject,
+      options?: boolean | AddEventListenerOptions
+    ) => any;
+    removeEventListener: Function;
+  },
+  M,
+  K extends keyof M
+>(
+  el: RefTyped<T>,
+  name: K,
+  listener: (this: T, ev: M[K]) => any,
+  options?: boolean | AddEventListenerOptions
+): RemoveEventFunction;
 export function useEvent<K extends keyof WindowEventMap>(
   el: RefTyped<Window>,
   name: K,
@@ -16,7 +48,7 @@ export function useEvent<K extends keyof DocumentEventMap>(
   options?: boolean | AddEventListenerOptions
 ): RemoveEventFunction;
 export function useEvent(
-  el: RefTyped<Element> | RefTyped<Window>,
+  el: RefTyped<Element> | RefTyped<Window> | RefTyped<any>,
   name: string,
   listener: EventListenerOrEventListenerObject,
   options?: boolean | AddEventListenerOptions