+++ packages/web/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./event";
 export * from "./web";
 export * from "./misc";
+export * from "./storage";